FEAT: implementa a camada de Serviço para Compra

+++ Mercado-Web-API/Data/Interface Repository/ICompraRepository.cs
@@ -4,9 +4,9 @@ namespace Mercado_Web_API.Data.Interface_Repository {
     public interface ICompraRepository {
         public void Add(Compra compra);
         public List<Compra> GetAll();
-        public Compra GetById(int id);
+        public Compra GetById(long id);
         public List<Compra> GetComprasByClienteId(int clienteId);
-        public List<Item> GetItensByCompraId(int compraId);
+        public List<Item> GetItensByCompraId(long compraId);
         public void Delete(Compra compra);
     }
 }
